New printMasterLayout example
This script will take an app, then create a human readable printout that contains every attribute name, internal ID, and the names/ids for any values of that field.

There are still some pending capabilities I've identified in the comments at the bottom of the script.

This is to be used as a tool to help when developing API scripts for your Apptivo account.